worker/requirements.txt
cffi==1.14.4
Pillow==8.0.1
pycparser==2.20
pyheif==0.5.1
